Revise back and implement forward button
Switched to a modulo-wrap implementation so we
are not trying to match up the clue texts to find
which clue is shown right now. Now, we simply store
active clue index.

Slight cost to storage efficiency by storing
direction and index in the clue objects, but worth
it for ease of use.